Knight Movement
Made three actions for the knights which are selected by random. AttackPlayer will simply move in the shortest path to the player. RandomisedLocation will make the knight move to a random location on the canvas and wait there. Decoy will make the knight go the limits of the y axis and then attack the player from above, allowing for the attackMove knights to have a frontal attack.